+++ src/main/java/me/superckl/prayers/capability/DefaultLivingPrayerUser.java
@@ -4,16 +4,16 @@ import net.minecraft.entity.LivingEntity;
 
 public class DefaultLivingPrayerUser extends LivingPrayerUser<LivingEntity>{
 
-	private final float maxPoints;
+	private final double maxPoints;
 
-	public DefaultLivingPrayerUser(final LivingEntity ref, final float maxPoints) {
+	public DefaultLivingPrayerUser(final LivingEntity ref, final double maxPoints) {
 		super(ref);
 		this.maxPoints = maxPoints;
 		this.currentPrayerPoints = maxPoints;
 	}
 
 	@Override
-	public float getMaxPrayerPoints() {
+	public double getMaxPrayerPoints() {
 		return this.maxPoints;
 	}
